Ok, I'm going to go about this a different way, so I reverted some code.
The plan is to at the moment, only until I find a way to actually do
some rtp streaming from a headless server where I can simply add files
to the rtp stream and have them played at the other end, do file copies.

All the audio will simply be generated or pre-existing files that are in
memory and get transferred and played.

The files should all be small enough that it doesnt cause too much of an
issue. We will see.